+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On a typical system during daylight hours, the Envoy will execute the task within 15 minutes.
 - More info about Power Production task here: https://support.enphase.com/s/article/How-do-I-disable-and-enable-power-production
 
+## [5.8.0] - (12.08.2022)
+## Changes
+- added possibility automatically 'Power peak reset' every day, week, month
+- config schema updated
+
 ## [5.7.8] - (08.08.2022)
 ## Changes
 - fix [#53](https://github.com/grzegorz914/homebridge-enphase-envoy/issues/53)